+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 - [11357](https://github.com/vegaprotocol/vega/issues/11357) - Support historical game scores
 - [11023](https://github.com/vegaprotocol/vega/issues/11023) - Add proposed fees to `vAMM` data.
 - [11028](https://github.com/vegaprotocol/vega/issues/11028) - Add API to estimate order book depth based on `vAMM`.
+- [11400](https://github.com/vegaprotocol/vega/issues/11400) - Add support for long block auction.
 
 ### 🐛 Fixes